+++ starknet_py/proxy/proxy_check.py
@@ -1,5 +1,5 @@
 from abc import ABC, abstractmethod
-from typing import List, Optional
+from typing import Optional
 
 from starkware.starknet.public.abi import (
     get_storage_var_address,
@@ -84,16 +84,3 @@ class OpenZeppelinProxyCheck(ProxyCheck):
             block_hash="latest",
         )
         return proxy_implementation_hash or None
-
-
-class ProxyResolutionError(Exception):
-    """
-    Error while resolving proxy using ProxyChecks.
-    """
-
-    def __init__(self, proxy_checks: List[ProxyCheck]):
-        proxy_checks_classes = [proxy_check.__class__ for proxy_check in proxy_checks]
-        self.message = (
-            f"Couldn't resolve proxy using given ProxyChecks ({proxy_checks_classes})"
-        )
-        super().__init__(self.message)
